Item rotation matches intended result.
Item rotates properly using the mouse position as the pivot point. The item widget and the validation squares match each other as intended. Issue remaining: When an item is rotated, the validation squares do not update until a new square is hovered over, so the function to update the validation squares need to be called upon rotation. Also, testing and further implementation is needed to make sure that the dragged item is clear on it use at the borders of the grid, in regards to item being out of bound.